+++ sunsu-wedding/src/main/java/com/kakao/sunsuwedding/user/planner/Planner.java
@@ -20,7 +20,7 @@ import java.time.LocalDateTime;
 public class Planner extends User {
 
     @Builder
-    public Planner(int id, String email, String password, String username, Grade grade, boolean is_active, LocalDateTime created_at, String order_id, int payed_amount, LocalDateTime payed_at) {
+    public Planner(Long id, String email, String password, String username, Grade grade, boolean is_active, LocalDateTime created_at, String order_id, Long payed_amount, LocalDateTime payed_at) {
         super(id, email, password, username, grade, is_active, created_at, order_id, payed_amount, payed_at);
     }
 }
